fix: update stale import paths after feed parser refactor
